fixed the variable and time definitions and how its broadcasted
